requirements.txt
Django==5.1
django-cors-headers==4.0.0
djangorestframework==3.15.2 
django-extensions==3.2.1
gunicorn==21.2.0
psycopg==3.2.1
python-dotenv==1.0.0
redis==5.2.1
celery==5.4.0 